Renamed folder from "contants" to "constants"
